Refactor: Consolidate backend entry points and improve project structure
