+++ organization_report.md
@@ -1,0 +1,98 @@
+# Persian Legal AI Project Organization Report
+
+**Generated on:** Thu Sep 11 07:29:36 AM UTC 2025
+**Script version:** 1.0
+**Total files processed:** 43
+
+## Summary
+
+- ✅ **Successfully moved:** 39 files
+- ❌ **Failed moves:** 0 files  
+- ⚠️ **Uncategorized:** 4 files
+
+## Directory Structure Created
+
+```
+docs/
+├── reports/          # All *REPORT*.md files
+├── guides/           # All *GUIDE*.md files  
+├── analysis/         # All analysis files
+└── archive/          # Old documentation
+
+tests/
+├── unit/            # Unit test files
+├── integration/     # Integration tests
+├── e2e/            # End-to-end tests
+└── fixtures/       # Test data
+
+scripts/
+├── deployment/     # Deployment scripts
+├── maintenance/    # Maintenance scripts
+└── utils/         # Utility scripts
+
+configs/
+├── development/    # Dev configs
+├── production/     # Prod configs
+└── testing/       # Test configs
+
+backup/
+├── databases/      # Database backups
+├── configs/        # Config backups
+└── logs/          # Log files
+```
+
+## Successfully Moved Files
+
+- BOLT_INTEGRATION_COMPLETE.md -> docs/reports/BOLT_INTEGRATION_COMPLETE.md
+- CLEANUP_SUMMARY.md -> docs/reports/CLEANUP_SUMMARY.md
+- MERGE_COMPLETION_REPORT.md -> docs/reports/MERGE_COMPLETION_REPORT.md
+- PERSIAN_AI_ENHANCEMENT_COMPLETION_REPORT.md -> docs/reports/PERSIAN_AI_ENHANCEMENT_COMPLETION_REPORT.md
+- PERSIAN_LEGAL_AI_AUDIT_COMPLETION_REPORT.md -> docs/reports/PERSIAN_LEGAL_AI_AUDIT_COMPLETION_REPORT.md
+- PHASE_3_COMPLETION_SUMMARY.md -> docs/reports/PHASE_3_COMPLETION_SUMMARY.md
+- PHASE_CRITICAL_PRODUCTION_ISSUES_RESOLVED.md -> docs/reports/PHASE_CRITICAL_PRODUCTION_ISSUES_RESOLVED.md
+- RECOVERY_REPORT.md -> docs/reports/RECOVERY_REPORT.md
+- SAFE_MERGE_COMPLETION_REPORT.md -> docs/reports/SAFE_MERGE_COMPLETION_REPORT.md
+- VISUAL_UI_TEST_REPORT.md -> docs/reports/VISUAL_UI_TEST_REPORT.md
+- integration-report.md -> docs/reports/integration-report.md
+- DEPLOYMENT_GUIDE.md -> docs/guides/DEPLOYMENT_GUIDE.md
+- START_SYSTEM.md -> docs/guides/START_SYSTEM.md
+- backend-implementation-guide.md -> docs/guides/backend-implementation-guide.md
+- BACKEND_MAIN_FILES_ANALYSIS.md -> docs/analysis/BACKEND_MAIN_FILES_ANALYSIS.md
+- analysis_report.txt -> docs/analysis/analysis_report.txt
+- test_frontend.html -> tests/e2e/test_frontend.html
+- test_integration.py -> tests/integration/test_integration.py
+- test_production_system.py -> tests/integration/test_production_system.py
+- system_health_check.py -> scripts/maintenance/system_health_check.py
+- start_system.py -> scripts/start_system.py
+- dependency-analyzer.js -> scripts/utils/dependency-analyzer.js
+- dependency-analyzer.ts -> scripts/utils/dependency-analyzer.ts
+- backend-validator-simple.js -> scripts/utils/backend-validator-simple.js
+- backend-validator.js -> scripts/utils/backend-validator.js
+- route-updater.js -> scripts/utils/route-updater.js
+- route-updater.ts -> scripts/utils/route-updater.ts
+- check-bolt-endpoints.sh -> scripts/check-bolt-endpoints.sh
+- comprehensive-test.sh -> scripts/comprehensive-test.sh
+- fix-imports.sh -> scripts/fix-imports.sh
+- fix-quotes.sh -> scripts/fix-quotes.sh
+- persian_legal_ai_recovery.sh -> scripts/maintenance/persian_legal_ai_recovery.sh
+- safe-merge.sh -> scripts/deployment/safe-merge.sh
+- safe-merge-to-main.sh -> scripts/deployment/safe-merge-to-main.sh
+- smart-analysis.sh -> scripts/smart-analysis.sh
+- smart-migration.sh -> scripts/smart-migration.sh
+- docker-compose.yml -> configs/development/docker-compose.yml
+- docker-compose.production.yml -> configs/production/docker-compose.production.yml
+- persian_legal_ai.db -> backup/databases/persian_legal_ai.db
+
+## Uncategorized Files
+
+- ORGANIZATION_SCRIPT_README.md
+- main.py
+- persian_main.py
+- project.json
+
+## Undo Instructions
+
+To undo this organization, run:
+```bash
+bash undo_organization.sh
+```
