+++ scripts/check-bolt-endpoints.sh
