+++ scripts/maintenance/persian_legal_ai_recovery.sh
