+++ scripts/smart-analysis.sh
