+++ scripts/utils/backend-validator-simple.js
