+++ scripts/utils/dependency-analyzer.js
