+++ scripts/utils/dependency-analyzer.ts
