+++ scripts/utils/route-updater.ts
