+++ undo_organization.sh
@@ -1,0 +1,43 @@
+mv 'docs/reports/BOLT_INTEGRATION_COMPLETE.md' 'BOLT_INTEGRATION_COMPLETE.md'
+mv 'docs/reports/CLEANUP_SUMMARY.md' 'CLEANUP_SUMMARY.md'
+mv 'docs/reports/MERGE_COMPLETION_REPORT.md' 'MERGE_COMPLETION_REPORT.md'
+mv 'docs/reports/PERSIAN_AI_ENHANCEMENT_COMPLETION_REPORT.md' 'PERSIAN_AI_ENHANCEMENT_COMPLETION_REPORT.md'
+mv 'docs/reports/PERSIAN_LEGAL_AI_AUDIT_COMPLETION_REPORT.md' 'PERSIAN_LEGAL_AI_AUDIT_COMPLETION_REPORT.md'
+mv 'docs/reports/PHASE_3_COMPLETION_SUMMARY.md' 'PHASE_3_COMPLETION_SUMMARY.md'
+mv 'docs/reports/PHASE_CRITICAL_PRODUCTION_ISSUES_RESOLVED.md' 'PHASE_CRITICAL_PRODUCTION_ISSUES_RESOLVED.md'
+mv 'docs/reports/RECOVERY_REPORT.md' 'RECOVERY_REPORT.md'
+mv 'docs/reports/SAFE_MERGE_COMPLETION_REPORT.md' 'SAFE_MERGE_COMPLETION_REPORT.md'
+mv 'docs/reports/VISUAL_UI_TEST_REPORT.md' 'VISUAL_UI_TEST_REPORT.md'
+mv 'docs/reports/integration-report.md' 'integration-report.md'
+mv 'docs/guides/DEPLOYMENT_GUIDE.md' 'DEPLOYMENT_GUIDE.md'
+mv 'docs/guides/START_SYSTEM.md' 'START_SYSTEM.md'
+mv 'docs/guides/backend-implementation-guide.md' 'backend-implementation-guide.md'
+mv 'docs/analysis/BACKEND_MAIN_FILES_ANALYSIS.md' 'BACKEND_MAIN_FILES_ANALYSIS.md'
+mv 'docs/analysis/analysis_report.txt' 'analysis_report.txt'
+mv 'tests/e2e/test_frontend.html' 'test_frontend.html'
+mv 'tests/integration/test_integration.py' 'test_integration.py'
+mv 'tests/integration/test_production_system.py' 'test_production_system.py'
+mv 'scripts/maintenance/system_health_check.py' 'system_health_check.py'
+mv 'scripts/start_system.py' 'start_system.py'
+mv 'scripts/utils/dependency-analyzer.js' 'dependency-analyzer.js'
+mv 'scripts/utils/dependency-analyzer.ts' 'dependency-analyzer.ts'
+mv 'scripts/utils/backend-validator-simple.js' 'backend-validator-simple.js'
+mv 'scripts/utils/backend-validator.js' 'backend-validator.js'
+mv 'scripts/utils/route-updater.js' 'route-updater.js'
+mv 'scripts/utils/route-updater.ts' 'route-updater.ts'
+mv 'scripts/check-bolt-endpoints.sh' 'check-bolt-endpoints.sh'
+mv 'scripts/comprehensive-test.sh' 'comprehensive-test.sh'
+mv 'scripts/fix-imports.sh' 'fix-imports.sh'
+mv 'scripts/fix-quotes.sh' 'fix-quotes.sh'
+mv 'scripts/maintenance/persian_legal_ai_recovery.sh' 'persian_legal_ai_recovery.sh'
+mv 'scripts/deployment/safe-merge.sh' 'safe-merge.sh'
+mv 'scripts/deployment/safe-merge-to-main.sh' 'safe-merge-to-main.sh'
+mv 'scripts/smart-analysis.sh' 'smart-analysis.sh'
+mv 'scripts/smart-migration.sh' 'smart-migration.sh'
+mv 'configs/development/docker-compose.yml' 'docker-compose.yml'
+mv 'configs/production/docker-compose.production.yml' 'docker-compose.production.yml'
+mv 'backup/databases/persian_legal_ai.db' 'persian_legal_ai.db'
+
+print_color "$GREEN" "✓ Undo completed successfully!"
+print_color "$YELLOW" "Note: Empty directories were not removed automatically."
+print_color "$YELLOW" "You may want to clean them up manually if needed."
